PandoraModel module killed
