+++ lib/pandora/constants.rb
@@ -35,5 +35,17 @@ module Pandora
     TI_Name  = 0
     TI_Type  = 1
     TI_Desc  = 2
+
+    # Panobject state flags
+    # RU: Флаги состояния объекта/записи
+    PSF_Support   = 1         # "поддерживаю", т.е. должна храниться
+    PSF_Hurvest   = 2         # собирается/загружается по частям
+    PSF_Trashed   = 0xFF00    # помечена к удалению
+    PSF_Hidden    = 0xFF01    # скрыто от показа в таблице
+
+    PK_Person  = 1
+    PK_Key     = 221
+    PK_Sign    = 222
+    PK_Message = 227
   end
 end
